move bark spec to prototype
