+++ torchaudio/functional/__init__.py
@@ -26,7 +26,6 @@ from .functional import (
     amplitude_to_DB,
     apply_beamforming,
     apply_codec,
-    barkscale_fbanks,
     compute_deltas,
     compute_kaldi_pitch,
     create_dct,
@@ -62,7 +61,6 @@ __all__ = [
     "compute_kaldi_pitch",
     "create_dct",
     "melscale_fbanks",
-    "barkscale_fbanks",
     "linear_fbanks",
     "DB_to_amplitude",
     "loudness",
