+++ torchaudio/functional/functional.py
@@ -22,7 +22,6 @@ __all__ = [
     "compute_deltas",
     "compute_kaldi_pitch",
     "melscale_fbanks",
-    "barkscale_fbanks",
     "linear_fbanks",
     "create_dct",
     "compute_deltas",
@@ -478,121 +477,6 @@ def _mel_to_hz(mels: Tensor, mel_scale: str = "htk") -> Tensor:
     freqs[log_t] = min_log_hz * torch.exp(logstep * (mels[log_t] - min_log_mel))
 
     return freqs
-
-
-def _hz_to_bark(freqs: float, bark_scale: str = "traunmuller") -> float:
-    r"""Convert Hz to Barks.
-
-    Args:
-        freqs (float): Frequencies in Hz
-        bark_scale (str, optional): Scale to use: ``traunmuller``, ``schroeder`` or ``wang``. (Default: ``traunmuller``)
-
-    Returns:
-        barks (float): Frequency in Barks
-    """
-
-    if bark_scale not in ["schroeder", "traunmuller", "wang"]:
-        raise ValueError('bark_scale should be one of "schroeder", "traunmuller" or "wang".')
-
-    if bark_scale == "wang":
-        return 6.0 * math.asinh(freqs / 600.0)
-    elif bark_scale == "schroeder":
-        return 7.0 * math.asinh(freqs / 650.0)
-    # Traunmuller Bark scale
-    barks = ((26.81 * freqs) / (1960.0 + freqs)) - 0.53
-    # Bark value correction
-    if barks < 2:
-        barks += 0.15 * (2 - barks)
-    elif barks > 20.1:
-        barks += 0.22 * (barks - 20.1)
-
-    return barks
-
-
-def _bark_to_hz(barks: Tensor, bark_scale: str = "traunmuller") -> Tensor:
-    """Convert bark bin numbers to frequencies.
-
-    Args:
-        barks (Tensor): Bark frequencies
-        bark_scale (str, optional): Scale to use: ``traunmuller``,``schroeder`` or ``wang``. (Default: ``traunmuller``)
-
-    Returns:
-        freqs (Tensor): Barks converted in Hz
-    """
-
-    if bark_scale not in ["schroeder", "traunmuller", "wang"]:
-        raise ValueError('bark_scale should be one of "traunmuller", "schroeder" or "wang".')
-
-    if bark_scale == "wang":
-        return 600.0 * torch.sinh(barks / 6.0)
-    elif bark_scale == "schroeder":
-        return 650.0 * torch.sinh(barks / 7.0)
-    # Bark value correction
-    if any(barks < 2):
-        idx = barks < 2
-        barks[idx] = (barks[idx] - 0.3) / 0.85
-    elif any(barks > 20.1):
-        idx = barks > 20.1
-        barks[idx] = (barks[idx] + 4.422) / 1.22
-
-    # Traunmuller Bark scale
-    freqs = 1960 * ((barks + 0.53) / (26.28 - barks))
-
-    return freqs
-
-
-def barkscale_fbanks(
-    n_freqs: int,
-    f_min: float,
-    f_max: float,
-    n_barks: int,
-    sample_rate: int,
-    bark_scale: str = "traunmuller",
-) -> Tensor:
-    r"""Create a frequency bin conversion matrix.
-
-    .. devices:: CPU
-
-    .. properties:: TorchScript
-
-    Args:
-        n_freqs (int): Number of frequencies to highlight/apply
-        f_min (float): Minimum frequency (Hz)
-        f_max (float): Maximum frequency (Hz)
-        n_barks (int): Number of mel filterbanks
-        sample_rate (int): Sample rate of the audio waveform
-        bark_scale (str, optional): Scale to use: ``traunmuller``,``schroeder`` or ``wang``. (Default: ``traunmuller``)
-
-    Returns:
-        Tensor: Triangular filter banks (fb matrix) of size (``n_freqs``, ``n_barks``)
-        meaning number of frequencies to highlight/apply to x the number of filterbanks.
-        Each column is a filterbank so that assuming there is a matrix A of
-        size (..., ``n_freqs``), the applied result would be
-        ``A * barkscale_fbanks(A.size(-1), ...)``.
-
-    """
-
-    # freq bins
-    all_freqs = torch.linspace(0, sample_rate // 2, n_freqs)
-
-    # calculate bark freq bins
-    m_min = _hz_to_bark(f_min, bark_scale=bark_scale)
-    m_max = _hz_to_bark(f_max, bark_scale=bark_scale)
-
-    m_pts = torch.linspace(m_min, m_max, n_barks + 2)
-    f_pts = _bark_to_hz(m_pts, bark_scale=bark_scale)
-
-    # create filterbank
-    fb = _create_triangular_filterbank(all_freqs, f_pts)
-
-    if (fb.max(dim=0).values == 0.0).any():
-        warnings.warn(
-            "At least one bark filterbank has all zero values. "
-            f"The value for `n_barks` ({n_barks}) may be set too high. "
-            f"Or, the value for `n_freqs` ({n_freqs}) may be set too low."
-        )
-
-    return fb
 
 
 def _create_triangular_filterbank(
