+++ torchaudio/transforms/__init__.py
@@ -1,13 +1,10 @@
 from ._multi_channel import MVDR, PSD, RTFMVDR, SoudenMVDR
 from ._transforms import (
     AmplitudeToDB,
-    BarkScale,
-    BarkSpectrogram,
     ComputeDeltas,
     Fade,
     FrequencyMasking,
     GriffinLim,
-    InverseBarkScale,
     InverseMelScale,
     InverseSpectrogram,
     LFCC,
@@ -37,16 +34,13 @@ __all__ = [
     "FrequencyMasking",
     "GriffinLim",
     "InverseMelScale",
-    "InverseBarkScale",
     "InverseSpectrogram",
     "LFCC",
     "Loudness",
     "MFCC",
     "MVDR",
     "MelScale",
-    "BarkScale",
     "MelSpectrogram",
-    "BarkSpectrogram",
     "MuLawDecoding",
     "MuLawEncoding",
     "PSD",
